# apps/api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Accord API
  version: 0.3.0
  description: >-
    Accord backend core service for simulated household negotiation.
    Privacy boundary: raw private interview transcripts are never exposed in shared API responses.
servers:
  - url: http://localhost:8000
paths:
...
  /api/households/{householdId}:
    get:
      summary: Get household
      parameters:
        - in: path
          name: householdId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Household details }
        '404': { description: Not found }
...
  /api/assignments/households/{householdId}:
    get:
      summary: List assignments for household
      parameters:
        - in: path
          name: householdId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Assignment list }
  /api/assignments/{assignmentId}/status:
    patch:
      summary: Update assignment status
      parameters:
        - in: path
          name: assignmentId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Status updated }
        '404': { description: Not found (including unauthorized) }
...
components:
  securitySchemes: {}